{{cookiecutter.project_name_dashes}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_name_dashes}}"
version = "{{cookiecutter.version}}"
description = "{{cookiecutter.description}}"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
        { include = "{{cookiecutter.project_name_underscores}}", from = "src" }
    ]

[tool.poetry.dependencies]
python = "^3.12"
cachetools = "^5.3.3"
fastapi = "^0.111.0"
pandas = "^2.2.2"
pydantic-settings = "^2.2.1"
setuptools = "^70.0.0"
uvicorn = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins."console_scripts"]
"{{cookiecutter.project_name_dashes}}" = "{{cookiecutter.project_name_underscores}}.cli:start_api"